passed tests: AR 18, 20; CO 16, 18, non-presidential 20; document load_and_test_all.py
