+++ docs/User_Guide.md
@@ -403,6 +403,10 @@ To unload existing data for a given jurisdiction and a given election -- or more
 
 where `test_dir` is the directory holding the tests to perform on the data before upload. For example, `test_dir` might be the repository's `tests` directory. This routine will move any files associated with unloaded data to the directory specified in the optional `unloaded_dir` in `run_time.ini`.
 
+## Testing
+The routine `tests/load_and_test_all.py` can be used to run tests. If the directory `test/TestingData` does not exist, the function will download files from `github.com/ElectionDataAnalysis/TestingData`, load it all and run all tests. If `tests/TestingData` exists, the routine will test all data in that directory (without downloading anything). Election-jurisdiction pairs can be specified with the -e and -j flags, e.g. `load_all_from_repo.py -e '2018 General' -j 'Arkansas'` to restrict the loading and testing to just that pair.
+
+
 ## Miscellaneous helpful hints
 Beware of:
  - Different names for same contest in different counties (if munging from a batch of county-level files)
