+++ tests/test_Arkansas_2018-General.py
@@ -39,9 +39,9 @@ def test_ar_house_totals_18(dbname):
 def test_ar_contest_by_vote_type_18(dbname):
     assert ( not e.data_exists("2018 General","Arkansas", dbname=dbname) or
             e.count_type_total(
-            "2016 General",
-            "North Carolina",
-"AR Senate District 5",
+            "2018 General",
+            "Arkansas",
+            "AR Senate District 5",
             "absentee",
             dbname=dbname,
         )
