+++ tests/test_Colorado_2016-General.py
@@ -1,5 +1,8 @@
 import election_data_analysis as e
 
+def test_data_exists(dbname):
+    assert e.data_exists("2016 General","Colorado",dbname=dbname)
+
 def test_co_presidential_16(dbname):
     assert (e.contest_total(
             "2016 General",
