+++ tests/test_dataloader.py
@@ -10,7 +10,7 @@ def test_ar_presidential_18(dbname):
 
 
 def test_mi_statewide_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Michigan",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Michigan",
             "US Senate MI (Democratic Party)",
@@ -21,7 +21,7 @@ def test_mi_statewide_totals_20(dbname):
 
 
 def test_mi_us_rep_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Michigan",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Michigan",
             "US House MI District 9 (Democratic Party)",
@@ -32,7 +32,7 @@ def test_mi_us_rep_totals_20(dbname):
 
 
 def test_mi_state_rep_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Michigan",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Michigan",
             "MI House District 37 (Republican Party)",
@@ -54,7 +54,7 @@ def test_mi_totals_match_vote_type_20(dbname):
 
 
 def test_mi_presidential_20ppp(dbname):
-    assert (not e.data_exists("2020 Primary","Michigan",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Michigan",
             "US President (MI) (Democratic Party)",
@@ -67,7 +67,7 @@ def test_mi_presidential_20ppp(dbname):
 ### Delaware 2020 Primary Data Loading Tests ###
 
 def test_de_statewide_totals(dbname):
-    assert (not e.data_exists("2020 Primary","Delaware",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Delaware",
                 "DE Governor (Republican Party)",
@@ -78,7 +78,7 @@ def test_de_statewide_totals(dbname):
 
 
 def test_de_senate_totals(dbname):
-    assert (not e.data_exists("2020 Primary","Delaware",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Delaware",
                 "DE Senate District 13 (Democratic Party)",
@@ -89,7 +89,7 @@ def test_de_senate_totals(dbname):
 
 
 def test_de_house_totals(dbname):
-    assert (not e.data_exists("2020 Primary","Delaware",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Delaware",
                 "DE House District 26 (Democratic Party)",
@@ -100,7 +100,7 @@ def test_de_house_totals(dbname):
 
 
 def test_de_contest_by_vote_type(dbname):
-    assert ( not e.data_exists("2020 Primary","Delaware", dbname=dbname) or
+    assert ( 
             e.count_type_total(
             "2016 General",
             "North Carolina",
@@ -113,7 +113,7 @@ def test_de_contest_by_vote_type(dbname):
 
 
 def test_de_presidential(dbname):
-    assert (not e.data_exists("2020 Primary","Delaware",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Delaware",
             "US President (DE) (Democratic Party)",
@@ -124,7 +124,7 @@ def test_de_presidential(dbname):
 
 
 def test_de_totals_match_vote_type(dbname):
-    assert (not e.data_exists("2020 Primary","Delaware", dbname=dbname) or 
+    assert ( 
             e.check_totals_match_vote_types("2020 Primary","Delaware" ,dbname=dbname) == True)
 
 
@@ -132,7 +132,7 @@ def test_de_totals_match_vote_type(dbname):
 ## oh 2016g tests
 
 def test_oh_presidential_16(dbname):
-    assert (not e.data_exists("2016 General","Ohio",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "Ohio",
             "US President (OH)",
@@ -148,7 +148,7 @@ def test_oh_statewide_16(dbname):
 
 
 def test_oh_senate_totals_16(dbname):
-    assert (not e.data_exists("2016 General","Ohio",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "Ohio",
             "OH Senate District 16",
@@ -160,7 +160,7 @@ def test_oh_senate_totals_16(dbname):
 
 
 def test_oh_house_totals_16(dbname):
-    assert (not e.data_exists("2016 General","Ohio",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "Ohio",
             "OH House District 2",
@@ -172,7 +172,7 @@ def test_oh_house_totals_16(dbname):
 
 
 def test_oh_contest_by_vote_type_16(dbname):
-    assert ( not e.data_exists("2016 General","Ohio", dbname=dbname) or
+    assert ( 
             e.count_type_total(
             "2016 General",
             "North Carolina",
@@ -186,7 +186,7 @@ def test_oh_contest_by_vote_type_16(dbname):
 
 
 def test_oh_totals_match_vote_type_16(dbname):
-    assert (not e.data_exists("2016 General","Ohio", dbname=dbname) or 
+    assert ( 
             e.check_totals_match_vote_types("2016 General","Ohio" ,dbname=dbname) == True)
 
 #OH18 test
@@ -196,7 +196,7 @@ def test_oh_presidential(dbname):
 
 
 def test_oh_statewide(dbname):
-    assert (not e.data_exists("2018 General","Ohio",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2018 General",
             "Ohio",
             "OH Governor",
@@ -207,7 +207,7 @@ def test_oh_statewide(dbname):
 
 
 def test_oh_senate_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Ohio",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2018 General",
             "Ohio",
             "OH Senate District 21",
@@ -219,7 +219,7 @@ def test_oh_senate_totals_18(dbname):
 
 
 def test_oh_house_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Ohio",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2018 General",
             "Ohio",
             "OH House District 2",
@@ -241,7 +241,7 @@ def test_oh_totals_match_vote_type_18(dbname):
 ### Illinois Data Loading Tests ###
 
 def test_il_presidential_16(dbname):
-    assert (not e.data_exists("2016 General","Illinois",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "Illinois",
             "US President (IL)",
@@ -252,7 +252,7 @@ def test_il_presidential_16(dbname):
 
 
 def test_il_statewide_totals_16(dbname):
-    assert (not e.data_exists("2016 General","Illinois",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "Illinois",
             "IL Comptroller",
@@ -263,7 +263,7 @@ def test_il_statewide_totals_16(dbname):
 
 
 def test_il_senate_totals_16(dbname):
-    assert (not e.data_exists("2016 General","Illinois",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "Illinois",
             "US Senate IL",
@@ -274,7 +274,7 @@ def test_il_senate_totals_16(dbname):
 
 
 def test_il_rep_16(dbname):
-    assert (not e.data_exists("2016 General","Illinois",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "Illinois",
             "IL Senate District 14",
@@ -298,7 +298,7 @@ def test_il_presidential_18(dbname):
 
 
 def test_il_statewide_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Illinois",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2018 General",
             "Illinois",
             "IL Governor",
@@ -313,7 +313,7 @@ def test_il_senate_totals_18(dbname):
 
 
 def test_il_rep_18(dbname):
-    assert (not e.data_exists("2018 General","Illinois",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2018 General",
             "Illinois",
             "IL House District 10",
@@ -333,7 +333,7 @@ def test_il_totals_match_vote_type_18(dbname):
 #IL20 test
 
 def test_il_presidential_20(dbname):
-    assert (not e.data_exists("2020 Primary","Illinois",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Illinois",
             "US President (IL)",
@@ -344,7 +344,7 @@ def test_il_presidential_20(dbname):
 
 
 def test_il_statewide_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Illinois",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Illinois",
             "US Senate IL",
@@ -355,7 +355,7 @@ def test_il_statewide_totals_20(dbname):
 
 
 def test_il_state_senate_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Illinois",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Illinois",
             "IL Senate District 11",
@@ -366,7 +366,7 @@ def test_il_state_senate_totals_20(dbname):
 
 
 def test_il_state_rep_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Illinois",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "Illinois",
             "IL House District 60",
@@ -388,7 +388,7 @@ def test_il_totals_match_vote_type_20(dbname):
 #CA16 test
 
 def test_ca_presidential_16(dbname):
-    assert (not e.data_exists("2016 General","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "California",
             "US President (CA)",
@@ -399,7 +399,7 @@ def test_ca_presidential_16(dbname):
 
 
 def test_ca_statewide_totals_16(dbname):
-    assert (not e.data_exists("2016 General","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "California",
             "US Senate CA",
@@ -410,7 +410,7 @@ def test_ca_statewide_totals_16(dbname):
 
 
 def test_ca_senate_totals_16(dbname):
-    assert (not e.data_exists("2016 General","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "California",
             "CA Senate District 15",
@@ -421,7 +421,7 @@ def test_ca_senate_totals_16(dbname):
 
 
 def test_ca_rep_16(dbname):
-    assert (not e.data_exists("2016 General","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2016 General",
             "California",
             "CA House District 60",
@@ -446,7 +446,7 @@ def test_ca_presidential_18(dbname):
 
 
 def test_ca_statewide_totals_18(dbname):
-    assert (not e.data_exists("2018 General","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2018 General",
             "California",
             "US Senate CA",
@@ -457,7 +457,7 @@ def test_ca_statewide_totals_18(dbname):
 
 
 def test_ca_senate_totals_18(dbname):
-    assert (not e.data_exists("2018 General","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2018 General",
             "California",
             "CA Senate District 12",
@@ -468,7 +468,7 @@ def test_ca_senate_totals_18(dbname):
 
 
 def test_ca_rep_18(dbname):
-    assert (not e.data_exists("2018 General","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2018 General",
             "California",
             "CA House District 60",
@@ -493,7 +493,7 @@ def test_ca_presidential_20(dbname):
 
 
 def test_ca_statewide_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "California",
             "US Senate CA",
@@ -504,7 +504,7 @@ def test_ca_statewide_totals_20(dbname):
 
 
 def test_ca_senate_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "California",
             "CA Senate District 12",
@@ -515,7 +515,7 @@ def test_ca_senate_totals_20(dbname):
 
 
 def test_ca_rep_18(dbname):
-    assert (not e.data_exists("2020 Primary","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 Primary",
             "California",
             "CA House District 60",
@@ -534,7 +534,7 @@ def test_ca_totals_match_vote_type_20(dbname):
 
 
 def test_ca_presidential_20ppp(dbname):
-    assert (not e.data_exists("2020 President Preference Primary","California",dbname=dbname) or e.contest_total(
+    assert ( e.contest_total(
             "2020 President Preference Primary",
             "California",
             "US President (CA)",
@@ -547,96 +547,6 @@ def test_ca_presidential_20ppp(dbname):
 ### Colorado Data Loading Tests ###
 
 
-#CO18 test
-
-def test_co_presidential_18(dbname):
-    assert True == True
-
-
-def test_co_statewide_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Colorado",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Colorado",
-            "CO Attorney General",
-            dbname=dbname,
-        )
-            == 2491954
-    )
-
-
-def test_co_senate_totals_18(dbname):
-    assert (not e.data_exists("2018 General","Colorado",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Colorado",
-            "CO Senate District 15",
-            dbname=dbname,
-        )
-            == 83690
-    )
-
-
-def test_co_rep_18(dbname):
-    assert (not e.data_exists("2018 General","Colorado",dbname=dbname) or e.contest_total(
-            "2018 General",
-            "Colorado",
-            "CO House District 60",
-            dbname=dbname,
-        )
-            == 39237
-    )
-
-
-def test_co_contest_by_vote_type_18(dbname):
-    assert True == True
-
-
-def test_co_totals_match_vote_type_18(dbname):
-    assert True == True
-
-#CO20 test
-
-def test_co_presidential_20(dbname):
-    assert (not e.data_exists("2020 Primary","Colorado",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Colorado",
-            "US President (CO) (Democratic Party)",
-            dbname=dbname,
-        )
-            == 960128
-    )
-
-
-def test_co_statewide_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Colorado",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Colorado",
-            "US Senate CO (Republican Party)",
-            dbname=dbname,
-        )
-            == 554806
-    )
-
-
-def test_co_senate_totals_20(dbname):
-    assert (not e.data_exists("2020 Primary","Colorado",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Colorado",
-            "CO Senate District 21 (Republican Party)",
-            dbname=dbname,
-        )
-            == 6320
-    )
-
-
-def test_co_rep_20(dbname):
-    assert (not e.data_exists("2020 Primary","Colorado",dbname=dbname) or e.contest_total(
-            "2020 Primary",
-            "Colorado",
-            "CO House District 20 (Democratic Party)",
-            dbname=dbname,
-        )
-            == 10011
-    )
 
 
 def test_co_contest_by_vote_type_18(dbname):
